+++ src/event.ts
@@ -8,8 +8,7 @@ import { IProject } from './project';
 import { IPropertyValue } from './propertyValue';
 import { SortType } from './sortType';
 
-// Pickで再定義(2023-08-07~)
-// export type IOffer = Omit<OfferFactory.IOffer, 'project' | 'priceCurrency'>;
+export type IUnacceptedPaymentMethodType = 'MovieTicket';
 export type IOffer = Pick<
     OfferFactory.IOffer,
     'typeOf' | 'unacceptedPaymentMethod' | 'seller' | 'eligibleQuantity' | 'itemOffered' | 'offeredThrough'
@@ -55,8 +54,6 @@ export interface IAttributes<T extends EventType> {
      * イベントがキャンセル、あるいは、延期された場合に主に使用されます。
      */
     eventStatus: EventStatusType;
-    // 廃止(2022-09-06~)
-    // hasOfferCatalog?: IHasOfferCatalog;
     headline?: IMultilingualString;
     /**
      * イベントが実行される場所
@@ -119,7 +116,7 @@ export interface ISearchConditions<T extends EventType> {
      * イベントタイプ
      */
     typeOf?: T;
-    // $inでの検索を可能にする)(2023-07-13~)
+    // $inでの検索を可能にする(2023-07-13~)
     typeOfIn?: EventType[];
     /**
      * イベント名称